Combine the changes into fewer files.
