--- src/main/scala/com/mwz/sonar/scala/pr/Issue.scala
@@ -1,31 +1,0 @@
-/*
- * Copyright (C) 2018-2019  All sonar-scala contributors
- *
- * This program is free software: you can redistribute it and/or modify
- * it under the terms of the GNU Lesser General Public License as published
- * by the Free Software Foundation, either version 3 of the License, or
- * (at your option) any later version.
- *
- * This program is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- * GNU Lesser General Public License for more details.
- *
- * You should have received a copy of the GNU General Lesser Public License
- * along with this program.  If not, see <http://www.gnu.org/licenses/>.
- */
-
-package com.mwz.sonar.scala
-package pr
-
-import org.sonar.api.batch.fs.InputFile
-import org.sonar.api.batch.rule.Severity
-import org.sonar.api.rule.RuleKey
-
-final case class Issue(
-  key: RuleKey,
-  file: InputFile,
-  line: Int,
-  severity: Severity,
-  message: String
-)
