+++ src/main/scala/com/mwz/sonar/scala/pr/github/Codec.scala
@@ -17,6 +17,7 @@
 
 package com.mwz.sonar.scala
 package pr
+package github
 
 import cats.Applicative
 import io.circe.generic.extras.Configuration
