+++ src/main/scala/com/mwz/sonar/scala/pr/github/Github.scala
@@ -17,9 +17,10 @@
 
 package com.mwz.sonar.scala
 package pr
+package github
 
 import cats.effect.Sync
-import com.mwz.sonar.scala.pr.Codec._
+import com.mwz.sonar.scala.pr.github.Codec._
 import io.circe.generic.auto._
 import org.http4s.client.Client
 import org.http4s.{Header, Headers, Method, Request, Uri}
