+++ src/main/scala/com/mwz/sonar/scala/pr/model.scala
@@ -18,8 +18,19 @@
 package com.mwz.sonar.scala
 package pr
 
-final case class ReviewStatus(blocker: Int, critical: Int)
+import org.http4s.Uri
 
+// TODO: Capture more custom errors.
+sealed trait ReviewError extends Exception
+case object NoFilesInPR extends ReviewError
+
+sealed trait PrStatus extends Product with Serializable
+case object Pending extends PrStatus
+case object Success extends PrStatus
+final case class Error(reviewStatus: ReviewStatus) extends PrStatus
+final case class Failure(error: Throwable) extends PrStatus
+
+final case class ReviewStatus(blocker: Int, critical: Int)
 object ReviewStatus {
   def description(reviewStatus: ReviewStatus): String = {
     reviewStatus match {
@@ -36,4 +47,20 @@ object ReviewStatus {
 
   private[pr] def form(i: Int): String =
     if (i > 1) "s" else ""
+}
+
+final case class Markdown(text: String) extends AnyVal
+object Markdown {
+
+  /**
+   * Generate an inline comment.
+   * The format is: "SEVERITY: TEXT ([more](link to the rule))"
+   */
+  def inline(baseUrl: Uri, issue: Issue): Markdown = {
+    val ruleUri: Uri = baseUrl
+      .withPath("coding_rules")
+      .withQueryParam("open", issue.key.toString)
+    // TODO: Add severity image.
+    Markdown(s"${issue.severity.name}: ${issue.message} ([more]($ruleUri))")
+  }
 }
